Use Rack::Sendfile instead of directly setting X-Accel-Redirect
This requires a change in the NGiNX configuration so that Rack::Sendfile is made aware that it should add the X-Accel-Redirect header and omit the file body. See: https://github.com/rack/rack/blob/1.6.0/lib/rack/sendfile.rb#L44-L45

The advantage to this is that it makes Doc-Juan work with any local dev server. If the proxy pass header is not present, Doc-Juan will instead simply serve up the file in the request body. This means that we get the same performance win in production of having NGiNX serve the file, but we do not have to do anything special in development mode.

Also, the code just reads nicer ;)